--- a/lib/ContentTypeParser.js
+++ b/lib/ContentTypeParser.js
@@ -43,16 +43,16 @@ ContentTypeParser.prototype.getHandler = function (contentType) {
   }
 }
 
-ContentTypeParser.prototype.run = function (contentType, handler, context, params, req, res, query) {
-  this.getHandler(contentType)(req, done)
-
+ContentTypeParser.prototype.run = function (contentType, requestContext, callback) {
   function done (body) {
     if (body instanceof Error) {
-      const reply = new context.Reply(res, context, null)
-      return reply.send(body)
+      callback(body, requestContext)
+      return
     }
-    handler(context, params, req, res, body, query)
+    callback(null, requestContext)
   }
+
+  this.getHandler(contentType)(requestContext.req, done)
 }
 
 function buildContentTypeParser (c) {