--- a/lib/reply.js
+++ b/lib/reply.js
@@ -20,61 +20,8 @@ function Reply (res, context, request) {
   this.request = request
 }
 
-/**
- * Instead of using directly res.end(), we are using setImmediate(…)
- * This because we have observed that with this technique we are faster at responding to the various requests,
- * since the setImmediate forwards the res.end at the end of the poll phase of libuv in the event loop.
- * So we can gather multiple requests and then handle all the replies in a different moment,
- * causing a general improvement of performances, ~+10%.
- */
 Reply.prototype.send = function (payload) {
-  if (this.sent) {
-    this.res.log.warn(new Error('Reply already sent'))
-    return
-  }
-
-  this.sent = true
-
-  var _isError = this[isError]
-
-  if (payload === undefined && _isError !== true) {
-    this.res.setHeader('Content-Length', '0')
-    if (!this.res.getHeader('Content-Type')) {
-      this.res.setHeader('Content-Type', 'text/plain')
-    }
-    setImmediate(handleReplyEnd, this, '')
-    return
-  }
-
-  if (payload && payload.isBoom) {
-    this.res.log.error(payload)
-    this.res.statusCode = payload.output.statusCode
-
-    this.res.setHeader('Content-Type', 'application/json')
-    this.headers(payload.output.headers)
-
-    setImmediate(
-      handleReplyEnd,
-      this,
-      payload.output.payload
-    )
-    return
-  }
-
-  if (payload instanceof Error || _isError === true) {
-    handleError(this, payload, wrapHandleReplyEnd)
-    return
-  }
-
-  if (payload && payload._readableState) {
-    if (!this.res.getHeader('Content-Type')) {
-      this.res.setHeader('Content-Type', 'application/octet-stream')
-    }
-    return pump(payload, this.res, wrapPumpCallback(this))
-  }
-
-  setImmediate(handleReplyEnd, this, payload)
-  return
+  setImmediate(onSendEnd, this, payload)
 }
 
 Reply.prototype.header = function (key, value) {
@@ -116,14 +63,6 @@ Reply.prototype.redirect = function (code, url) {
   }
 
   this.header('Location', url).code(code).send()
-}
-
-function wrapHandleReplyEnd (reply, payload) {
-  setImmediate(
-    handleReplyEnd,
-    reply,
-    payload
-  )
 }
 
 function wrapPumpCallback (reply) {