--- a/lib/reply.js
+++ b/lib/reply.js
@@ -117,13 +117,18 @@ function onSendEnd (reply, payload) {
   }
 
   if (!reply.res.getHeader('Content-Length')) {
-    reply.res.setHeader('Content-Length', '' + Buffer.byteLength(payload))
+    if (payload !== null && payload !== undefined) {
+      reply.res.setHeader('Content-Length', '' + Buffer.byteLength(payload))
+    } else {
+      reply.res.setHeader('Content-Length', '0')
+    }
   }
+
   reply.sent = true
   reply.res.end(payload)
 }
 
-function handleError (reply, err, cb) {
+function handleError (reply, err) {
   if (!reply.res.statusCode || reply.res.statusCode < 400) {
     reply.res.statusCode =
         (err === null || err === undefined) ? 500
@@ -149,11 +154,15 @@ function handleError (reply, err, cb) {
 
   reply.res.setHeader('Content-Type', 'application/json')
 
-  cb(reply, Object.assign({
+  reply.send(Object.assign({
     error: statusCodes[reply.res.statusCode + ''],
     message: (err || '') && err.message,
     statusCode: reply.res.statusCode
   }, reply._extendServerError && reply._extendServerError(err)))
+}
+
+Reply.prototype.sendError = function (error) {
+  handleError(this, error)
 }
 
 function buildReply (R) {