--- a/package.json
+++ b/package.json
@@ -102,6 +102,7 @@
     "abstract-logging": "^1.0.0",
     "ajv": "^5.3.0",
     "avvio": "^3.2.0",
+    "fast-fast-series": "^0.1.0",
     "fast-iterator": "^0.2.1",
     "fast-json-stringify": "^0.15.1",
     "fastify-cli": "^0.11.0",