@jsumners 2017-11-30
It will take me some time to really understand this enough to add a proper review. In the interim I'd like to offer a GitHub tip: when you are looking at a file you can press <kbd>y</kbd> to adjust the URL bar to the commit you are viewing. Thereby creating a permanent link to the source code in question. The way you have linked your "this" examples currently, the URLs will be invalid at some point in the future.