@allevo 2017-12-01
> There are too many failing test, I think that is not fair run benchmarks between the two implementations since you proposal supports a very small part of our api.

I agree. The benchmarks could be fake. But the aim of this implementation is to not add overhead if some "hook" are added.

> I fear that your solution (fast-fast-series) will slow down the hooks execution, I'd like to see a benchmark there. 

I'll provide some bench obviously.