@allevo 2017-12-01
@delvedor benchmark done https://gist.github.com/allevo/a862c2490e779b95e480ec15c0da3293 . I'd like to move this discussion there