@allevo 2017-12-23
Hi guys,
  I'm sorry but for the next days (and the previous ones) I have no time for implementing this PR.

Currently there're a lot of PRs welcomed in `fastify` and it is hard for me to keep this PR up to date. So I'm closing it without deleting the branch.

Keep thinking outside the box! 🤘